% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImbTreeAUC.R
\name{PrintTree}
\alias{PrintTree}
\title{Print an ImbTreeAUC Object}
\usage{
PrintTree( tree )
}
\arguments{
\item{tree}{Fitted model object. This is assumed to be the result of some function that produces an object with the same named components as that returned by the ImbTreeAUC function.}
}
\value{
Printed an ImbTreeAUC based model.
}
\seealso{
  \code{\link{ImbTreeAUC}}, \code{\link{ImbTreeAUCInter}}, \code{\link{PredictTree}}, \code{\link{PrintTree}}, \code{\link{PrintTreeInter}}, \code{\link{ExtractRules}}
}

\examples{
library("ImbTreeAUC")
data(iris)
Tree <- ImbTreeAUC( Y_name = "Species", 
                    X_names = colnames(iris)[-ncol(iris)], 
                    data = iris)
}